+++ src/FORMAT_VALUE_TYPE.mjs
@@ -1,0 +1,5 @@
+export const FORMAT_VALUE_TYPE_COLOR = "color";
+
+export const FORMAT_VALUE_TYPE_EMAIL = "email";
+
+export const FORMAT_VALUE_TYPE_URL = "url";