+++ src/FluxValueFormat.mjs
@@ -1,4 +1,4 @@
-import { VALUE_FORMAT_TYPE_COLOR, VALUE_FORMAT_TYPE_EMAIL, VALUE_FORMAT_TYPE_TEXT, VALUE_FORMAT_TYPE_URL } from "./VALUE_FORMAT_TYPE.mjs";
+import { DEFAULT_FORMAT_VALUE_TYPES } from "./DEFAULT_FORMAT_VALUE_TYPES.mjs";
 
 /** @typedef {import("./formatValue.mjs").formatValue} formatValue */
 
@@ -6,7 +6,7 @@ export class FluxValueFormat {
     /**
      * @type {Map<string, formatValue>}
      */
-    #formats;
+    #format_values;
 
     /**
      * @returns {FluxValueFormat}
@@ -19,7 +19,17 @@ export class FluxValueFormat {
      * @private
      */
     constructor() {
-        this.#formats = new Map();
+        this.#format_values = new Map();
+
+        for (const [
+            type,
+            format_value
+        ] of Object.entries(DEFAULT_FORMAT_VALUE_TYPES)) {
+            this.addFormatValue(
+                type,
+                format_value
+            );
+        }
     }
 
     /**
@@ -27,108 +37,37 @@ export class FluxValueFormat {
      * @param {formatValue} format_value
      * @returns {void}
      */
-    addFormat(type, format_value) {
-        this.#formats.set(type, format_value);
+    addFormatValue(type, format_value) {
+        if (this.#format_values.has(type)) {
+            throw new Error(`Format value type ${type} already exists`);
+        }
+
+        this.#format_values.set(type, format_value);
     }
 
     /**
      * @param {*} value
      * @param {string | null} type
-     * @returns {Promise<string | Node>}
+     * @returns {Promise<Node | string>}
      */
     async formatValue(value = null, type = null) {
-        switch (type ?? VALUE_FORMAT_TYPE_TEXT) {
-            case VALUE_FORMAT_TYPE_COLOR:
-                if ((value ?? "") !== "") {
-                    return (await import("./Color/FluxValueFormatColorElement.mjs")).FluxValueFormatColorElement.new(
-                        value
-                    );
-                } else {
-                    return this.formatValue(
-                        value,
-                        VALUE_FORMAT_TYPE_TEXT
-                    );
-                }
+        let formatted_value;
 
-            case VALUE_FORMAT_TYPE_EMAIL:
-                if (value === null || typeof value !== "object") {
-                    return this.formatValue(
-                        {
-                            email: value
-                        },
-                        VALUE_FORMAT_TYPE_EMAIL
-                    );
-                }
+        if ((type ?? "") !== "") {
+            const format_value = this.#format_values.get(type) ?? null;
 
-                if ((value.email ?? "") !== "") {
-                    return (await import("./Url/FluxValueFormatUrlElement.mjs")).FluxValueFormatUrlElement.new(
-                        `mailto:${value.email}`,
-                        (value.label ?? "") !== "" ? value.label : value.email,
-                        value.title ?? null
-                    );
-                } else {
-                    return this.formatValue(
-                        value.email,
-                        VALUE_FORMAT_TYPE_TEXT
-                    );
-                }
-
-            case VALUE_FORMAT_TYPE_TEXT:
-                return `${(value ?? "") !== "" ? value : "-"}`;
-
-            case VALUE_FORMAT_TYPE_URL:
-                if (value === null || typeof value !== "object") {
-                    return this.formatValue(
-                        {
-                            url: value
-                        },
-                        VALUE_FORMAT_TYPE_URL
-                    );
-                }
-
-                if ((value.url ?? "") !== "") {
-                    return (await import("./Url/FluxValueFormatUrlElement.mjs")).FluxValueFormatUrlElement.new(
-                        value.url,
-                        value.label ?? null,
-                        value.title ?? null
-                    );
-                } else {
-                    return this.formatValue(
-                        value.url,
-                        VALUE_FORMAT_TYPE_TEXT
-                    );
-                }
-
-            default: {
-                let format_value;
-
-                if (!this.#formats.has(type) || (format_value = this.#formats.get(type) ?? null) === null) {
-                    throw new Error(`Unknown format type ${type}`);
-                }
-
-                return format_value(
-                    value
-                );
+            if (format_value === null) {
+                throw new Error(`Unknown format value type ${type}`);
             }
-        }
-    }
 
-    /**
-     * @param {HTMLElement} element
-     * @param {*} value
-     * @param {string | null} type
-     * @returns {Promise<void>}
-     */
-    async formatValueToElement(element, value = null, type = null) {
-        const formatted_value = await this.formatValue(
-            value,
-            type
-        );
+            formatted_value = await format_value(
+                value
+            ) ?? null;
 
-        if (formatted_value instanceof Node) {
-            element.appendChild(formatted_value);
         } else {
-            element.innerText = formatted_value;
+            formatted_value = value;
         }
+
+        return typeof Node !== "undefined" && formatted_value instanceof Node ? formatted_value : `${(formatted_value ?? "") !== "" ? formatted_value : "-"}`;
     }
 }