+++ src/Url/FluxFormatUrlValueElement.mjs
@@ -1,21 +1,21 @@
 import { flux_css_api } from "../../../flux-css-api/src/FluxCssApi.mjs";
 
 const variables_css = await flux_css_api.import(
-    `${import.meta.url.substring(0, import.meta.url.lastIndexOf("/"))}/FluxValueFormatUrlElementVariables.css`
+    `${import.meta.url.substring(0, import.meta.url.lastIndexOf("/"))}/FluxFormatUrlValueElementVariables.css`
 );
 
 document.adoptedStyleSheets.unshift(variables_css);
 
 const css = await flux_css_api.import(
-    `${import.meta.url.substring(0, import.meta.url.lastIndexOf("/"))}/FluxValueFormatUrlElement.css`
+    `${import.meta.url.substring(0, import.meta.url.lastIndexOf("/"))}/FluxFormatUrlValueElement.css`
 );
 
-export class FluxValueFormatUrlElement extends HTMLElement {
+export class FluxFormatUrlValueElement extends HTMLElement {
     /**
      * @param {string} url
      * @param {string | null} label
      * @param {string | null} title
-     * @returns {FluxValueFormatUrlElement}
+     * @returns {FluxFormatUrlValueElement}
      */
     static new(url, label = null, title = null) {
         return new this(
@@ -52,6 +52,6 @@ export class FluxValueFormatUrlElement extends HTMLElement {
     }
 }
 
-export const FLUX_VALUE_FORMAT_URL_ELEMENT_TAG_NAME = "flux-value-format-url";
+export const FLUX_FORMAT_URL_VALUE_ELEMENT_TAG_NAME = "flux-format-url-value";
 
-customElements.define(FLUX_VALUE_FORMAT_URL_ELEMENT_TAG_NAME, FluxValueFormatUrlElement);
+customElements.define(FLUX_FORMAT_URL_VALUE_ELEMENT_TAG_NAME, FluxFormatUrlValueElement);