+++ src/VALUE_FORMAT_TYPE.mjs
@@ -1,7 +1,0 @@
-export const VALUE_FORMAT_TYPE_COLOR = "color";
-
-export const VALUE_FORMAT_TYPE_EMAIL = "email";
-
-export const VALUE_FORMAT_TYPE_TEXT = "text";
-
-export const VALUE_FORMAT_TYPE_URL = "url";