+++ src/formatValueType.mjs
@@ -1,0 +1,3 @@
+/**
+ * @typedef {(value?: *, type?: string | null) => Promise<Node | string>} formatValueType
+ */