+++ CHANGELOG.md
@@ -6,6 +6,13 @@ Changes:
 
 \-
 
+## v2023-06-01-1
+
+Changes:
+
+- Split default format values
+- Remove `FORMAT_VALUE_TYPE_TEXT` and `formatValueToElement`
+
 ## v2023-05-30-1
 
 Changes: